+++ packages/dart/lib/src/objects/parse_file_base.dart
@@ -48,4 +48,8 @@ abstract class ParseFileBase extends ParseObject {
   Future<ParseResponse> upload({ProgressCallback? progressCallback});
 
   Future<ParseFileBase> download({ProgressCallback? progressCallback});
+
+  void progressCallback(ProgressCallback progressCallback);
+
+  void cancel([dynamic reason]);
 }